+++ zebra-network/src/peer_set/set/tests.rs
@@ -1,9 +1,6 @@
 use std::{net::SocketAddr, sync::Arc};
 
-use futures::{
-    channel::{mpsc, oneshot},
-    stream, Stream, StreamExt,
-};
+use futures::{channel::mpsc, stream, Stream, StreamExt};
 use proptest::{collection::vec, prelude::*};
 use proptest_derive::Arbitrary;
 use tokio::{sync::broadcast, task::JoinHandle};
@@ -21,10 +18,7 @@ use zebra_chain::{
 
 use super::MorePeers;
 use crate::{
-    peer::{
-        CancelHeartbeatTask, Client, ClientRequest, ErrorSlot, LoadTrackedClient,
-        MinimumPeerVersion,
-    },
+    peer::{LoadTrackedClient, MinimumPeerVersion, MockedClientHandle},
     peer_set::PeerSet,
     protocol::external::{types::Version, InventoryHash},
     AddressBook, Config,
@@ -37,50 +31,6 @@ mod prop;
 ///
 /// This affects the maximum number of peer connections added to the [`PeerSet`] during the tests.
 const MAX_PEERS: usize = 20;
-
-/// A handle to a mocked [`Client`] instance.
-struct MockedClientHandle {
-    _request_receiver: mpsc::Receiver<ClientRequest>,
-    shutdown_receiver: oneshot::Receiver<CancelHeartbeatTask>,
-    version: Version,
-}
-
-impl MockedClientHandle {
-    /// Create a new mocked [`Client`] instance, returning it together with a handle to track it.
-    pub fn new(version: Version) -> (Self, LoadTrackedClient) {
-        let (shutdown_sender, shutdown_receiver) = oneshot::channel();
-        let (request_sender, _request_receiver) = mpsc::channel(1);
-
-        let client = Client {
-            shutdown_tx: Some(shutdown_sender),
-            server_tx: request_sender,
-            error_slot: ErrorSlot::default(),
-            version,
-        };
-
-        let handle = MockedClientHandle {
-            _request_receiver,
-            shutdown_receiver,
-            version,
-        };
-
-        (handle, client.into())
-    }
-
-    /// Gets the peer protocol version associated to the [`Client`].
-    pub fn version(&self) -> Version {
-        self.version
-    }
-
-    /// Checks if the [`Client`] instance has not been dropped, which would have disconnected from
-    /// the peer.
-    pub fn is_connected(&mut self) -> bool {
-        match self.shutdown_receiver.try_recv() {
-            Ok(None) => true,
-            Ok(Some(CancelHeartbeatTask)) | Err(oneshot::Canceled) => false,
-        }
-    }
-}
 
 /// A helper type to generate arbitrary peer versions which can then become mock peer services.
 #[derive(Arbitrary, Debug)]
