+++ zebra-network/src/peer_set/set/tests/prop.rs
@@ -6,11 +6,9 @@ use tower::{discover::Discover, BoxError, ServiceExt};
 
 use zebra_chain::{block, chain_tip::ChainTip, parameters::Network};
 
-use super::{
-    BlockHeightPairAcrossNetworkUpgrades, MockedClientHandle, PeerSetBuilder, PeerVersions,
-};
+use super::{BlockHeightPairAcrossNetworkUpgrades, PeerSetBuilder, PeerVersions};
 use crate::{
-    peer::{LoadTrackedClient, MinimumPeerVersion},
+    peer::{LoadTrackedClient, MinimumPeerVersion, MockedClientHandle},
     peer_set::PeerSet,
     protocol::external::types::Version,
 };
